requirements.txt
aiogram==3.22.0
python-dotenv==1.0.0

asyncpg==0.29.0
redis[hiredis]==5.0.1